feat: add filter component
- Creates a dedicated component for the transaction filter

- Separates the filter logic from the Home page for better organization
