+++ src/pages/Home/index.jsx
@@ -2,36 +2,16 @@ import { useEffect } from "react";
 import { useState } from "react";
 import ModalAddTransaction from "../../components/ModalAddTransaction";
 import ListTransactions from "../../components/ListTransactions";
+import Filter from "../../components/Filter";
 
 function Home() {
   const [transactions, setTransactions] = useState([]);
 
-  const monthName = new Date()
-    .toLocaleString("en-US", { month: "long" })
-    .toLowerCase();
-
-  const [currentMonth, setCurrentMonth] = useState(monthName);
-
   const [queryParams, setQueryParams] = useState({
     limit: 99,
     query_type: "and",
-    month: currentMonth,
+    month: new Date().toLocaleString("en-US", { month: "long" }).toLowerCase(),
   });
-
-  const months = [
-    "january",
-    "february",
-    "march",
-    "april",
-    "may",
-    "june",
-    "july",
-    "august",
-    "september",
-    "october",
-    "november",
-    "december",
-  ];
 
   useEffect(() => {
     getTransactions();
@@ -89,34 +69,11 @@ function Home() {
             </div>
           </div>
 
-          <section className="mt-4 overflow-x-scroll ">
-            <div className="flex py-4 justify-start">
-              {months.map((month, key) => {
-                return (
-                  <div
-                    key={key}
-                    className={`
-                    ${key == 0 ? "mr-2" : "mx-2"}
-                    lg:mx-auto 
-                    p-2 
-                    ${currentMonth == month ? "bg-blue-400" : "bg-blue-300"}
-                    text-white 
-                    text-sm 
-                    min-w-20 
-                    text-center
-                    capitalize
-                  `}
-                    onClick={() => {
-                      setCurrentMonth(month);
-                      setQueryParams({ ...queryParams, month: month });
-                    }}
-                  >
-                    {month}
-                  </div>
-                );
-              })}
-            </div>
-          </section>
+          <Filter
+            handleSelectedMonth={(month) =>
+              setQueryParams({ ...queryParams, month: month })
+            }
+          />
 
           <ListTransactions
             transactions={transactions}
